dont update metdata when http paused
