+++ sdk/src/main/java/io/customer/sdk/queue/Queue.kt
@@ -94,7 +94,6 @@ class QueueImpl internal constructor(
         } else {
             // created a coroutine scope on UI thread because Android's CountdownTimer in SimpleTimer class must be
             // created and started on the UI thread or it will crash the SDK.
-            // TODO refactor this code to exist in the SimpleTimer class for easy encapsulation
             val job = CoroutineScope(uiDispatcher).launch {
                 // Not enough tasks in the queue yet to run it now, so let's schedule them to run in the future.
                 // It's expected that only 1 timer instance exists and is running in the SDK.
@@ -106,8 +105,6 @@ class QueueImpl internal constructor(
 
                 if (didSchedule) logger.info("queue timer: scheduled to run queue in $numberSecondsToScheduleTimer seconds")
             }
-
-            // TODO save job and cancel it when you run queueTimer.cancel()
         }
     }
 }
