+++ sdk/src/main/java/io/customer/sdk/queue/QueueRunRequest.kt
@@ -1,5 +1,6 @@
 package io.customer.sdk.queue
 
+import io.customer.sdk.error.CustomerIOError
 import io.customer.sdk.queue.type.QueueInventory
 import io.customer.sdk.queue.type.QueueTaskMetadata
 import io.customer.sdk.util.Logger
@@ -56,15 +57,20 @@ class QueueRunRequestImpl internal constructor(
                 val error = result.exceptionOrNull()
                 logger.debug("queue task $nextTaskStorageId run failed $error")
 
-                // TODO implement pauses in HTTP requests
-                // TODO parse the error to see if it was because of paused HTTP requests
-                val previousRunResults = nextTaskToRun.runResults
-                val newRunResults = nextTaskToRun.runResults.copy(totalRuns = previousRunResults.totalRuns + 1)
+                val customerIOError = error as? CustomerIOError
+                return if (customerIOError is CustomerIOError.HttpRequestsPaused) {
+                    // When HTTP requests are paused, don't increment metadata about the tasks to create inaccurate data
+                    logger.info("queue is quitting early because all HTTP requests are paused.")
 
-                logger.debug("queue task $nextTaskStorageId, updating run history from: $previousRunResults to: $newRunResults")
-                queueStorage.update(nextTaskStorageId, newRunResults)
+                    goToNextTask(emptyList(), totalNumberOfTasksToRun, null)
+                } else {
+                    val previousRunResults = nextTaskToRun.runResults
+                    val newRunResults = nextTaskToRun.runResults.copy(totalRuns = previousRunResults.totalRuns + 1)
+                    logger.debug("queue task $nextTaskStorageId, updating run history from: $previousRunResults to: $newRunResults")
+                    queueStorage.update(nextTaskStorageId, newRunResults)
 
-                return goToNextTask(inventory, totalNumberOfTasksToRun, nextTaskToRunInventoryItem)
+                    goToNextTask(inventory, totalNumberOfTasksToRun, nextTaskToRunInventoryItem)
+                }
             }
         }
     }
